Move prepared::Text::wrap_lines fn to sub-module
